Put all classes in Echobot package
