--- src/main/java/command/CommandResponse.java
@@ -1,22 +1,0 @@
-package command;
-
-public class CommandResponse {
-    private boolean isExit = false;
-    private String response;
-
-    public CommandResponse(boolean isExitCommand) {
-        this.isExit = isExitCommand;
-    }
-
-    public CommandResponse(String response) {
-        this.response = response;
-    }
-
-    public String getResponse() {
-        return this.response;
-    }
-
-    public boolean isExitCommand() {
-        return this.isExit;
-    }
-}
