--- src/main/java/command/ExitCommand.java
@@ -1,12 +1,0 @@
-package command;
-
-import exception.EchoBotException;
-
-public class ExitCommand extends Command {
-    public final static String COMMAND = "bye";
-
-    @Override
-    public CommandResponse execute() throws EchoBotException {
-        return new CommandResponse(true);
-    }
-}
