--- src/main/java/command/ListAllCommand.java
@@ -1,10 +1,0 @@
-package command;
-
-import exception.EchoBotException;
-
-public class ListAllCommand extends ListCommand {
-    @Override
-    public CommandResponse execute() throws EchoBotException {
-        return super.getTaskListResponse(taskList.getTaskList());
-    }
-}
