+++ src/main/java/echobot/exception/DeadlineEmptyException.java
@@ -1,0 +1,11 @@
+package echobot.exception;
+
+public class DeadlineEmptyException extends EchoBotException {
+    public DeadlineEmptyException() {
+        super("Deadline cannot be left empty!");
+    }
+
+    public DeadlineEmptyException(String message) {
+        super(message);
+    }
+}
