+++ src/main/java/echobot/exception/EchoBotException.java
@@ -1,0 +1,11 @@
+package echobot.exception;
+
+public class EchoBotException extends Exception {
+    public EchoBotException() {
+        super();
+    }
+
+    public EchoBotException(String message) {
+        super(message);
+    }
+}
