+++ src/main/java/echobot/exception/EventStartEndDateEmptyException.java
@@ -1,0 +1,11 @@
+package echobot.exception;
+
+public class EventStartEndDateEmptyException extends EchoBotException {
+    public EventStartEndDateEmptyException() {
+        super("task.Event start date and end date must both be stated!!!");
+    }
+
+    public EventStartEndDateEmptyException(String message) {
+        super(message);
+    }
+}
