+++ src/main/java/echobot/exception/InvalidDeadlineFormatException.java
@@ -1,0 +1,11 @@
+package echobot.exception;
+
+public class InvalidDeadlineFormatException extends EchoBotException {
+    public InvalidDeadlineFormatException() {
+        super("Deadline format is not valid!");
+    }
+
+    public InvalidDeadlineFormatException(String message) {
+        super("Invalid deadline format! Please follow " + message + "!");
+    }
+}
