+++ src/main/java/echobot/exception/TaskNameEmptyException.java
@@ -1,0 +1,11 @@
+package echobot.exception;
+
+public class TaskNameEmptyException extends EchoBotException {
+    public TaskNameEmptyException() {
+        super("OOPS!!! The task name could not be left empty!!!");
+    }
+
+    public TaskNameEmptyException(String message) {
+        super(message);
+    }
+}
