+++ src/main/java/echobot/exception/TaskNotFoundException.java
@@ -1,0 +1,11 @@
+package echobot.exception;
+
+public class TaskNotFoundException extends EchoBotException {
+    public TaskNotFoundException() {
+        super("Cannot find this task in the list!");
+    }
+
+    public TaskNotFoundException(String message) {
+        super(message);
+    }
+}
