+++ src/main/java/echobot/io/Saveable.java
@@ -1,0 +1,5 @@
+package echobot.io;
+
+public interface Saveable {
+    String save();
+}
