+++ src/main/java/echobot/task/ToDo.java
@@ -1,0 +1,19 @@
+package echobot.task;
+
+import echobot.exception.TaskNameEmptyException;
+
+public class ToDo extends Task {
+    public ToDo(Boolean isDone, String taskName) throws TaskNameEmptyException {
+        super(isDone, taskName);
+    }
+
+    @Override
+    public String save() {
+        return "T " + super.save();
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
