--- src/main/java/exception/EchoBotException.java
@@ -1,11 +1,0 @@
-package exception;
-
-public class EchoBotException extends Exception {
-    public EchoBotException() {
-        super();
-    }
-
-    public EchoBotException(String message) {
-        super(message);
-    }
-}
