--- src/main/java/exception/TaskNotFoundException.java
@@ -1,11 +1,0 @@
-package exception;
-
-public class TaskNotFoundException extends EchoBotException {
-    public TaskNotFoundException() {
-        super("Cannot find this task in the list!");
-    }
-
-    public TaskNotFoundException(String message) {
-        super(message);
-    }
-}
