--- src/main/java/io/Saveable.java
@@ -1,5 +1,0 @@
-package io;
-
-public interface Saveable {
-    String save();
-}
